[FE] 사전질문 관리 메뉴 분리 (#678)
* fix: 사전질문 관리 메뉴 분리

* fix: 사전질문 버튼 크기 수정
